Give the buffer traits one home, in binius-utils
BufferData and VecLike are container vocabulary — a DerefMut<Target = [T]>
that can shrink, and the subset of Vec's API callers rely on. Neither says
anything about compute, pooling, or fields, so move both to a new
binius_utils::buffer module and leave binius-compute holding only Allocator
and its two PoolVec impls.

The orphan rule decides the split: the Vec and &mut [T] impls name foreign
types, so they travel with the trait definition, while PoolVec's impls stay
with PoolVec.
