+++ crates/ip-prover/src/fracaddcheck/mod.rs
@@ -2,7 +2,7 @@
 
 use std::iter;
 
-use binius_compute::{Allocator, VecLike};
+use binius_compute::Allocator;
 use binius_field::{Field, PackedField};
 use binius_ip::{
 	fracaddcheck::FracAddEvalClaim, mlecheck, prodcheck::MultilinearEvalClaim,
@@ -14,9 +14,12 @@ use binius_math::{
 	line::extrapolate_line,
 	multilinear::eq::{eq_ind_partial_eval, eq_one_var},
 };
-use binius_utils::rayon::{
-	iter::{IntoParallelIterator, IntoParallelRefMutIterator, ParallelIterator},
-	task_size::{IndexedParallelIteratorExt, WorkPerItem},
+use binius_utils::{
+	buffer::VecLike,
+	rayon::{
+		iter::{IntoParallelIterator, IntoParallelRefMutIterator, ParallelIterator},
+		task_size::{IndexedParallelIteratorExt, WorkPerItem},
+	},
 };
 use either::Either;
 use itertools::izip;
